+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Add option to write packets to file as PCAP Capture File Format.
+
 ## [0.1.0] - 2024-11-05
 
 - Initial release